Added menu for deleting entries & detaching images
Added a menu on the add and view pages to delete an entry or detach an image from an entry.

Attaching images has been moved from the FAB to the image card so the FAB can be used for saving on the image page.

Added ImageManager class to tidy up MainActivity.